chore(api): add resource modules
