+++ moon.yml
@@ -28,3 +28,9 @@ tasks:
     command: "cargo fmt --all-targets"
   lint:
     command: "cargo clippy --all-targets -- --deny warnings"
+  docbuild:
+    command: "nix build .#kickbase-api-doc"
+  docs:
+    deps:
+      - "docbuild"
+    command: "$BROWSER result/share/index.html"
